Refactor tests: split to unit and integration test suites (RAD-745)
