+++ phpunit.xml.dist
@@ -5,8 +5,11 @@
     colors="true"
 >
     <testsuites>
-        <testsuite name="Unit tests">
-            <directory>./tests</directory>
+        <testsuite name="unit">
+            <directory>./tests/unit</directory>
+        </testsuite>
+        <testsuite name="integration">
+            <directory>./tests/integration</directory>
         </testsuite>
     </testsuites>
