+++ tests/unit/Http/Plugin/ExceptionPluginTest.php
