+++ tests/unit/Model/Command/InteractionTest.php
