+++ tests/unit/Model/Command/ItemPropertySetupTest.php
