+++ tests/unit/Model/Command/ItemPropertyTest.php
